Fix constituency / region data in the UK 2016 import script
The script used to create the elections, posts, areas, etc. for the UK
2016 regional assembly elections unfortunately mixed up the mapit_code
and the slug / name / party_lists_in_use for Scotland and Wales.

These fixes to the import script will mean that no one running that
script for the first time has that problem, so this is part of the fix
for #600.