CR: add a script to create posts, elections, area types, etc.
Note that this assumes that there is a single organization that every
councillor is elected to ('Consejo Municipal') whereas actually there
are many local Consejo Municipal rather than one single one. This
doesn't actually make a difference to the operation of the site, but
it's not correct. (The same is true of mayors and the Alcaldía
Municipal.)