Quick fix to suppress the honorific prefix / suffix for Costa Rica
We're going to switch to defining the Popolo fields that are available
for display / editing in the database, which would be a better way
of suppressing these two fields for Costa Rica than the conditionals
in the template used here.  (Accesa have told us that these fields are
of no use to them, so we can assume they're empty and hidden from the
start.)

For the moment, just do this quick fix so that the site is displayed
as they want.  Hopefully we'll do the change to make field selection
more flexible soon.